refactor(ui): extract ChatSidebar from ChatView
ChatView.tsx was 3429 LOC with the session list, search, rename
state, and date-grouping helpers all inlined alongside the
composer, transcript, header, and modals. The sidebar is the
cleanest cut: self-contained state (search query, renaming-id,
hover-id, scroll ref) plus a handful of pure helpers that only
the sidebar uses.

ChatSidebar owns its local state and pulls runtime context
directly via useRuntimeConsoleContext(). Two callbacks flow
through props for cross-region coordination:
  - onSelectSession(id): ChatView wires this to clear the draft,
    focus the composer textarea, and dispatch selectChatSession.
    Keeping the focus side-effect on the parent avoids the
    sidebar reaching across the canvas for the composer ref.
  - onCreateChat(): ChatView opens the draft, closes the chat
    settings panel, focuses the composer, and dispatches
    createChatSession. The sidebar gates this on
    newChatAgentReady internally.

Helpers moved into ChatSidebar.tsx:
  - filterSidebarSessions, groupSidebarSessions, sidebarDateGroup
    (sidebar-only)
  - sidebarSessionBrand, sidebarSessionAgentLabel,
    sidebarSessionTimeLabel (also used by ChatView's header /
    derived state — re-imported from ChatSidebar)
  - SidebarSession type

Composer extraction is deferred. The composer's derived-value
graph (composerVisible, composerRepair, hecateChatModelReady,
sendDisabled, agentBusy, queueingMessage, selectableModels,
hecateProviderOptions, hecateDisabledProviderReasons,
selectedCapabilityProvider, …) is read by ChatView's body for
header / modals / transcript as well, so moving it cleanly needs
either a shared derived-state hook or a much larger prop bag than
fits in one surgical PR. Splitting it off here keeps this one
under the stop-condition LOC budget.

Net diff:
  - ChatView.tsx: 3429 → 3130 LOC (-299)
  - ChatSidebar.tsx: new, 335 LOC
  - Combined: +36 LOC, but ChatView is meaningfully more focused.

Verification:
  - bun run typecheck — clean (tsgo -b)
  - bunx vitest run — 824/824 pass (38 files, no count delta)
  - bun run test:e2e — 60/60 chromium pass
  - Manual: dev server in Tauri webview, sidebar renders with the
    same buttons / search / empty state; New Hecate chat button
    + adapter dropdown + rename / delete hover icons all wired.
